feat: move type creation util to database module
